+++ src/main/java/com/cafe/website/payload/AreaDTO.java
@@ -12,7 +12,7 @@ public class AreaDTO extends BaseEntityDTO implements HasImageDTO {
 		super();
 	}
 
-	public AreaDTO(int id, Integer status, String createdAt, String updatedAt, String name, String slug,
+	public AreaDTO(Long id, Integer status, String createdAt, String updatedAt, String name, String slug,
 			ImageDTO image) {
 		super(id, status, createdAt, updatedAt);
 		this.name = name;
