+++ src/main/java/com/cafe/website/payload/ProductDeleteDTO.java
@@ -4,11 +4,11 @@ import jakarta.validation.constraints.NotNull;
 
 public class ProductDeleteDTO {
 	@NotNull
-	Integer productId;
+	Long productId;
 	@NotNull
-	Integer userId;
+	Long userId;
 
-	public ProductDeleteDTO(Integer productId, Integer userId) {
+	public ProductDeleteDTO(Long productId, Long userId) {
 		super();
 		this.productId = productId;
 		this.userId = userId;
@@ -17,19 +17,19 @@ public class ProductDeleteDTO {
 	public ProductDeleteDTO() {
 	}
 
-	public Integer getProductId() {
+	public Long getProductId() {
 		return productId;
 	}
 
-	public void setProductId(Integer productId) {
+	public void setProductId(Long productId) {
 		this.productId = productId;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
