+++ src/main/java/com/cafe/website/payload/ReviewCreateDTO.java
@@ -11,9 +11,9 @@ import jakarta.validation.constraints.NotNull;
 
 public class ReviewCreateDTO {
 	@NotNull
-	private Integer userId;
+	private Long userId;
 	@NotNull
-	private Integer productId;
+	private Long productId;
 
 	@NotNull
 	private List<MultipartFile> listImageFiles = new ArrayList<>();
@@ -33,7 +33,7 @@ public class ReviewCreateDTO {
 	@JsonIgnoreProperties(ignoreUnknown = true)
 	private List<FileMetadata> listFileMetadatas = new ArrayList<>();
 
-	public ReviewCreateDTO(@NotNull Integer userId, @NotNull Integer productId,
+	public ReviewCreateDTO(@NotNull Long userId, @NotNull Long productId,
 			@NotNull List<MultipartFile> listImageFiles, @NotNull String name, @NotNull int location,
 			@NotNull int space, @NotNull int food, @NotNull int service, @NotNull int price,
 			List<FileMetadata> listFileMetadatas) {
@@ -50,19 +50,19 @@ public class ReviewCreateDTO {
 		this.listFileMetadatas = listFileMetadatas;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
 
-	public Integer getProductId() {
+	public Long getProductId() {
 		return productId;
 	}
 
-	public void setProductId(Integer productId) {
+	public void setProductId(Long productId) {
 		this.productId = productId;
 	}
