+++ src/main/java/com/cafe/website/repository/AreaRepository.java
@@ -20,7 +20,7 @@ import jakarta.persistence.criteria.Order;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 
-public interface AreaRepository extends JpaRepository<Area, Integer> {
+public interface AreaRepository extends JpaRepository<Area, Long> {
 
 	@Query
 	default List<Area> findWithFilters(Integer status, String name, String slug, String createdAt, String updatedAt,
@@ -62,13 +62,13 @@ public interface AreaRepository extends JpaRepository<Area, Integer> {
 				.setMaxResults(pageable.getPageSize()).getResultList();
 	}
 
-	Boolean existsBySlugAndIdNot(String slug, Integer id);
+	Boolean existsBySlugAndIdNot(String slug, Long id);
 
 	Boolean existsBySlug(String slug);
 
 	Boolean existsByName(String name);
 
-	Boolean existsByNameAndIdNot(String name, Integer id);
+	Boolean existsByNameAndIdNot(String name, Long id);
 
 	Optional<Area> findByName(String name);
